+++ examples/render_single_template.rs
@@ -1,16 +1,17 @@
 use tera::Context;
+
 use terarium::{Content, Template, TerariumBuilder};
 
 /// The Terarium can render single template.
 fn main() {
-    let terarium = TerariumBuilder::default()
-        .add_template(
-            "my_template".to_owned(),
-            Template::default()
-                .add_content(Content::new("This is my template #{{tpl_number}}".to_owned(), vec!["en".to_owned()])).unwrap()
-                .add_content(Content::new("Toto je šablona #{{tpl_number}}".to_owned(), vec!["cs".to_owned()])).unwrap()
-        ).unwrap()
-        .build().unwrap();
+    let mut tpl = Template::default();
+    tpl.add_content(Content::new("This is my template #{{tpl_number}}".to_owned(), vec!["en".to_owned()])).unwrap();
+    tpl.add_content(Content::new("Toto je šablona #{{tpl_number}}".to_owned(), vec!["cs".to_owned()])).unwrap();
+
+    let mut builder = TerariumBuilder::default();
+    builder.add_template("my_template".to_owned(), tpl).unwrap();
+
+    let terarium = builder.build().unwrap();
 
     let mut ctx = Context::new();
     ctx.insert("tpl_number", "13");